Updating AddSearchPaths and SquirrelAwareExecutable to be 0 to many
